fix: scholership with dev url
