+++ apps/front-end/public/locales/en/translation.json
@@ -113,6 +113,7 @@
   "APPLY_FOR_LEAVE": "Apply For Leave",
   "APPLY_NOW": "Apply Now",
   "APPLY_SCHOLARSHIP": "Apply scholarship",
+  "SCHOLARSHIP_SUBMITTED_SUCCESSFULLY": "Scholarship has been submitted successfully",
   "APPROVE_APPLICATION": "Approve Application",
   "ARE_YOU_SURE": "Are you sure",
   "AS_PER_AADHAAR": "(As Per Aadhaar Card)",
