+++ apps/front-end/public/locales/hi/translation.json
@@ -112,6 +112,7 @@
   "APPLY_FOR_LEAVE": "छुट्टी के लिए आवेदन करें",
   "APPLY_NOW": "अभी अप्लाई करें",
   "APPLY_SCHOLARSHIP": "स्कालर्शिप अप्लाई करें",
+  "SCHOLARSHIP_SUBMITTED_SUCCESSFULLY": "स्कालर्शिप सफलतापूर्वक सबमिट कर दिया गया है",
   "APPROVE_APPLICATION": "आवेदन स्वीकृत करें",
   "ARE_YOU_SURE": "क्या आप निश्चित है",
   "AS_PER_AADHAAR": "(आधार कार्ड के अनुसार)",
