+++ lib/common-lib/src/components/admin_layout/Sidebar.js
@@ -37,7 +37,11 @@ export const Menus = ({ menus, routeDynamics, ...props }) => {
   const navigate = useNavigate()
   useEffect(() => {
     let path = window?.location?.pathname.toString()
-    if (path.startsWith('/admin/facilitator')) {
+    const arrData = menus.map((e) => e?.route)
+    const index = arrData.indexOf(path.split('/').slice(0, 3).join('/'))
+    if (index > 0) {
+      setSelected(index)
+    } else if (path.startsWith('/admin/facilitator')) {
       setSelected(1)
     } else if (path.startsWith('/admin/learners')) {
       setSelected(2)
